backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP client
aiohttp==3.9.3
httpx==0.26.0

# DNS resolution
dnspython==2.5.0

# Async file operations
aiofiles==23.2.1

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Templating
jinja2==3.1.3

# Task queue
celery==5.3.6

# Logging
python-json-logger==2.0.7

# Development
python-dotenv==1.0.1

# Rate limiting
slowapi==0.1.9